Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -2,9 +2,9 @@
 float main()
 {
     float a,b;
-    scanf("%f %f",&a,&b);
+    scanf("%.2f %f",&a,&b);
     
-    int product = a*b;
+    float product = a*b;
     printf("Product: %.2f\n", product);
     
     return 0;
